// src/api/generated/models/create-question-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ankieter API
 * Ankieter 6.0 - the API contract, source of truth for the backend and the frontend
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { QuestionPurposeDto } from './question-purpose-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionSourceDto } from './question-source-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionTypeDto } from './question-type-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionVersionContentDto } from './question-version-content-dto';

/**
 * 
 * @export
 * @interface CreateQuestionRequestDto
 */
export interface CreateQuestionRequestDto {
    /**
     * Leave empty to have one issued from the category prefix and a counter.
     * @type {string}
     * @memberof CreateQuestionRequestDto
     */
    'businessKey'?: string;
    /**
     * 
     * @type {QuestionTypeDto}
     * @memberof CreateQuestionRequestDto
     */
    'type': QuestionTypeDto;
    /**
     * 
     * @type {QuestionPurposeDto}
     * @memberof CreateQuestionRequestDto
     */
    'purpose': QuestionPurposeDto;
    /**
     * 
     * @type {string}
     * @memberof CreateQuestionRequestDto
     */
    'categoryId': string;
    /**
     * 
     * @type {QuestionSourceDto}
     * @memberof CreateQuestionRequestDto
     */
    'source': QuestionSourceDto;
    /**
     * 
     * @type {string}
     * @memberof CreateQuestionRequestDto
     */
    'sourceName'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateQuestionRequestDto
     */
    'tagIds'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateQuestionRequestDto
     */
    'positionCodes'?: Array<string>;
    /**
     * 
     * @type {QuestionVersionContentDto}
     * @memberof CreateQuestionRequestDto
     */
    'version': QuestionVersionContentDto;
}